Moved CampatibilityInfoBinaryPackge into new internal folder
Removed packages.config
