--- source/ClientDebug/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Serilog" version="2.12.0" targetFramework="net472" />
-  <package id="Serilog.Sinks.Console" version="4.1.0" targetFramework="net472" />
-  <package id="Serilog.Sinks.File" version="5.0.0" targetFramework="net472" />
-</packages>
